Experiment with alternative means of scene asm generation
Using a stateful builder to generate rather than using the stateless
generator class. This let's each event configure the scene ASM
arbitrarily, rather than just adding a single block of some ASM. It also
avoids the reliance on type checks to know what to do with the resulting
ASM which is otherwise ambiguous. Using a true Visitor also decouples
the model from the generator. The driver was that DisplayText needed to
be able to configure not just scene event ASM OR dialog ASM, but both.

It is also worth noting that this more cleanly encapsultes generation
context.