Add remaining portraits
This prompted changing the model a bit. It's still a bit weird and
probably deserves further changes.

The necessity of some of the complexity right now comes from a shortcut
in the google doc generator, which refers to the assembly sprite label
directly of a map object in order to display a "portrait" which is
infact a sprite icon of sorts.

This is really a quirk of our doc format that the NPC is shown even
though they don't have a portrait, in dialog.