Add special handling for guild receptionist sprite
I guess they cut down on art costs by only drawing front and back for
the sprite, and perhaps programming costs by reusing the routine and
mappings for a common shopkeeper routine. In turn, this sprite, which
has fewer facings than other sprites, needed to be duplicated in vram at
a certain position so that the 'up' facing lines up with the right spot
in vram.