Finish both branches in interaction mode IfFlag
Fixes extra dialog window when last event is one which could run in dialog or event.

Previously we were finishing queued events in whatever mode the generation was last in, which would erroneously add events to the dialog loop if we ended in dialog. Now each branch is finished with common logic.